// plugins/home/src/components/CustomHomepage/types.ts
const RSJFTypeSchema: z.ZodType<RJSFSchema> = z.any();
const RSJFTypeUiSchema: z.ZodType<UiSchema> = z.any();
const ReactElementSchema: z.ZodType<ReactElement> = z.any();
// ...
export const WidgetSchema = z.object({
  name: z.string(),
  title: z.string().optional(),
  description: z.string().optional(),
  component: ReactElementSchema,
  width: z.number().positive('width must be positive number').optional(),
  height: z.number().positive('height must be positive number').optional(),
  minWidth: z.number().positive('minWidth must be positive number').optional(),
  maxWidth: z.number().positive('maxWidth must be positive number').optional(),
  minHeight: z
    .number()
    .positive('minHeight must be positive number')
    .optional(),
  maxHeight: z
    .number()
    .positive('maxHeight must be positive number')
    .optional(),
  settingsSchema: RSJFTypeSchema.optional(),
  uiSchema: RSJFTypeUiSchema.optional(),
  movable: z.boolean().optional(),
  deletable: z.boolean().optional(),
  resizable: z.boolean().optional(),
});

export type Widget = z.infer<typeof WidgetSchema>;